requirements.txt
imagepypelines
astropy
numpy
scipy
Pillow
